Crate ic_transport_types

Types related to the HTTP transport for the Internet Computer. Primarily used through ic-agent.

Modules

signed

Types representing signed messages.

Structs

Delegation

A delegation from one key to another.

Envelope

The authentication envelope, containing the contents and their signature.

InvalidRejectCodeError

Error returned from RejectCode::try_from.

NodeSignature

A response signature from an individual node.

ReadStateResponse

The response from a request to the read_state endpoint.

RejectResponse

An IC execution error received from the replica.

ReplyResponse

A successful reply to a canister call.

RequestId

A Request ID.

SignedDelegation

A Delegation that has been signed by an Identity.

SubnetMetrics

A list of subnet metrics.

Enums

CallResponse

The response from a request to the call endpoint.

EnvelopeContent

The content of an IC ingress message, not including any signature information.

QueryResponse

Possible responses to a query call.

RejectCode

See the interface spec.

RequestIdError

An error during the calculation of the RequestId.

RequestStatusResponse

The response of /api/v2/canister/<effective_canister_id>/read_state with request_status request type.

TransportCallResponse

The parsed response from a request to the v3 call endpoint. A request to the call endpoint.

Functions

to_request_id

Derive the request ID from a serializable data structure. This does not include the ic-request domain prefix.